research/requirements.txt
psycopg2-binary>=2.9
pandas>=2.0
plotly>=5.0
python-dotenv>=1.0
